Started dealing with path.py
The only method that should be specific to path.py
is its length and some condition on its constructor
ensuring that the paths are not created with three
ends or three heads. All the rest, like start
node or end node depend on the type of edges the
path has and they should be implemented apart.