An important API change that breaks backwards compatibilty
Basically the V, E methods now output a frozenset instead
of a dictionary. This has a lot of implications in the
overall structure of the API. But it reduced most of
the operations to set operations which bumped the performance
by 1/3 in unittests.
The new signature for V, E properties are inline with
the mathematical definition of a graph, which contains
a vertex set and an edge set.